+++ src/twg2/collections/buffers/ByteBufferPool.java
@@ -15,7 +15,7 @@ public class ByteBufferPool {
 	private int maxInstances;
 	private int capacity;
 	private boolean direct;
-	private List<ByteBuffer> buffers;
+	private List<ByteBuffer> pool;
 	private BitSet inUse;
 
 
@@ -29,7 +29,7 @@ public class ByteBufferPool {
 		this.maxInstances = maxBuffers;
 		this.capacity = bufferSize;
 		this.direct = direct;
-		this.buffers = new ArrayList<ByteBuffer>();
+		this.pool = new ArrayList<ByteBuffer>();
 		this.inUse = new BitSet();
 	}
 
@@ -38,8 +38,8 @@ public class ByteBufferPool {
 	 * @param instance the instance to return
 	 */
 	public boolean returnInstance(ByteBuffer instance) {
-		synchronized(this.buffers) {
-			int index = this.buffers.indexOf(instance);
+		synchronized(this.pool) {
+			int index = this.pool.indexOf(instance);
 			if(index > -1) {
 				this.inUse.set(index, false);
 				instance.clear();
@@ -54,7 +54,7 @@ public class ByteBufferPool {
 	 * @return true if any buffers are available, false if not
 	 */
 	public boolean hasInstance() {
-		return this.maxInstances == -1 || this.buffers.size() < this.maxInstances || this.findUnusedBuffer() > -1;
+		return this.maxInstances == -1 || this.pool.size() < this.maxInstances || this.findUnusedInstance() > -1;
 	}
 
 
@@ -63,15 +63,15 @@ public class ByteBufferPool {
 	 * @throws Exception if there is are no available instance and the maximum pool size has been reached
 	 */
 	public ByteBuffer getInstance() throws Exception {
-		synchronized(this.buffers) {
-			int unusedBufferIdx = this.findUnusedBuffer();
+		synchronized(this.pool) {
+			int unusedBufferIdx = this.findUnusedInstance();
 			if(unusedBufferIdx > -1) {
 				inUse.set(unusedBufferIdx, true);
-				return buffers.get(unusedBufferIdx);
+				return pool.get(unusedBufferIdx);
 			}
 
 			// If there is no more room for a new instance, throw an error
-			if(this.maxInstances > 0 && this.buffers.size() >= this.maxInstances) {
+			if(this.maxInstances > 0 && this.pool.size() >= this.maxInstances) {
 				throw new SizeLimitExceededException("Pool instance limit reached");
 			}
 
@@ -83,16 +83,16 @@ public class ByteBufferPool {
 			else {
 				newBuffer = ByteBuffer.allocate(capacity);
 			}
-			this.inUse.set(this.buffers.size(), true);
-			this.buffers.add(newBuffer);
+			this.inUse.set(this.pool.size(), true);
+			this.pool.add(newBuffer);
 			return newBuffer;
 		}
 	}
 
 
-	private int findUnusedBuffer() {
+	private int findUnusedInstance() {
 		// Search for an unused/returned buffer
-		int size = buffers.size();
+		int size = pool.size();
 		for(int i = 0; i < size; i++) {
 			if(inUse.get(i) == false) {
 				return i;